+++ src/AppserverIo/Appserver/Core/Api/Node/ExtractorNode.php
@@ -101,7 +101,7 @@ class ExtractorNode extends AbstractNode implements ExtractorNodeInterface
      * Returns the nodes primary key, the name by default.
      *
      * @return string The nodes primary key
-     * @see \AppserverIo\Appserver\Core\Api\Node\AbstractNode::getPrimaryKey()
+     * @see \AppserverIo\Description\Api\Node\AbstractNode::getPrimaryKey()
      */
     public function getPrimaryKey()
     {