+++ src/AppserverIo/Appserver/Core/Api/Node/ParamNode.php
@@ -117,7 +117,7 @@ class ParamNode extends AbstractValueNode
      * Returns the nodes primary key, the name by default.
      *
      * @return string The nodes primary key
-     * @see \AppserverIo\Appserver\Core\Api\Node\AbstractNode::getPrimaryKey()
+     * @see \AppserverIo\Description\Api\Node\AbstractNode::getPrimaryKey()
      */
     public function getPrimaryKey()
     {