+++ src/AppserverIo/Appserver/Core/Api/Node/StorageNode.php
@@ -71,7 +71,7 @@ class StorageNode extends AbstractNode
      * Returns the nodes primary key, the name by default.
      *
      * @return string The nodes primary key
-     * @see \AppserverIo\Appserver\Core\Api\Node\AbstractNode::getPrimaryKey()
+     * @see \AppserverIo\Description\Api\Node\AbstractNode::getPrimaryKey()
      */
     public function getPrimaryKey()
     {